% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_enrollment.R
\name{get_enrollment}
\alias{get_enrollment}
\title{Get Enrollment}
\usage{

  get_enrollment(link = "http://web.williams.edu/admin/registrar//catalog/catalog0910.pdf")
}
\arguments{
\item{link}{A link to a yearly pdf file on the Williams College Registrar's website.}
}
\value{
A numeric vector containing enrollment by year, from graduate, and then senior to freshman.
}
\description{
Gets the enrollment by year(at the beginning of the academic year) from an archived Williams Course Catalog PDF link
}

