% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_enrollment_data.R
\name{get_enrollment_data}
\alias{get_enrollment_data}
\title{Get Enrollment Data}
\usage{
get_enrollment_data()
}
\value{
A data frame containing enrollment
}
\description{
Gets a data frame of enrollment(by class) over the years at Williams College
}

