% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_geographical_distribution.R
\name{get_geographical_distribution}
\alias{get_geographical_distribution}
\title{Get Geographical Distribution}
\usage{

  get_geographical_distribution(link = "http://web.williams.edu/admin/registrar//catalog/catalog0910.pdf")
}
\arguments{
\item{link}{A link to a yearly pdf file on the Williams College Registrar's website.}
}
\value{
A list of size two with two dataframes(one domestic, one international) with two columns:state/country and number.
}
\description{
Gets the geographical distribution data from a archived Williams Course Catalog PDF link.
}

