% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_price.R
\name{get_price}
\alias{get_price}
\title{Get Price}
\usage{

  get_price(link = "http://web.williams.edu/admin/registrar//catalog/catalog0910.pdf")
}
\arguments{
\item{link}{A link to a yearly pdf file on the Williams College Registrar's website.}
}
\value{
Price of tuition for that academic year
}
\description{
Gets the tuition price from an archived Williams Course Catalog PDF link
}

