% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_price_data.R
\name{get_price_data}
\alias{get_price_data}
\title{Get Price Data}
\usage{
get_price_data()
}
\value{
A data frame containing prices
}
\description{
Gets a data frame of tuition prices over the years at Williams College
}

